some improvement in backend and frontend

+++ profileService/internal/domain/academicYear/entity.go
@@ -1,6 +1,6 @@
 package academicYear
 
 type AcademicYear struct {
-	YearID int
+	YearID int64
 	Name   string
 }
