+++ profileService/internal/domain/lab/entity.go
@@ -1,7 +1,7 @@
 package lab
 
 type Lab struct {
-	LabID       int
+	LabID       int64
 	Name        string
-	InstituteID int
+	InstituteID int64
 }
