+++ profileService/internal/domain/position/entity.go
@@ -1,6 +1,6 @@
 package position
 
 type Position struct {
-	PositionID int
+	PositionID int64
 	Name       string
 }
