+++ profileService/internal/domain/profileVersion/repository.go
@@ -4,7 +4,7 @@ import "context"
 
 type Repository interface {
 	AddProfileVersion(ctx context.Context, profile *ProfileVersion) error
-	GetVersionByProfileID(ctx context.Context, profileID int64, year int) (*ProfileVersion, error)
+	GetVersionByProfileID(ctx context.Context, profileID int64, year int64) (*ProfileVersion, error)
 	GetVersionByVersionID(ctx context.Context, profileID int64) (*ProfileVersion, error)
 }
