+++ profileService/internal/domain/programCourseInstance/entity.go
@@ -1,7 +1,7 @@
 package programcourseinstance
 
 type ProgramCourseInstance struct {
-	ProgramCourseID  int
-	ProgramID        int
-	CourseInstanceID int
+	ProgramCourseID  int64
+	ProgramID        int64
+	CourseInstanceID int64
 }
