+++ profileService/internal/domain/workload/repository.go
@@ -3,7 +3,7 @@ package workload
 import "context"
 
 type Repository interface {
-	GetSemesterWorkloadByVersionID(ctx context.Context, profileVersionID int64, semesterID int) (*Workload, error)
+	GetSemesterWorkloadByVersionID(ctx context.Context, profileVersionID int64, semesterID int64) (*Workload, error)
 	AddSemesterWorkload(ctx context.Context, workload *Workload) error
 	UpdateSemesterWorkload(ctx context.Context, workload *Workload) error
 }
