+++ profileService/internal/domain/workload/service.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Service interface {
-	GetSemesterWorkloadByVersionID(ctx context.Context, profileVersionID int64, semesterID int) (*Workload, error)
+	GetSemesterWorkloadByVersionID(ctx context.Context, profileVersionID int64, semesterID int64) (*Workload, error)
 	AddSemesterWorkload(ctx context.Context, workload *Workload) error
 	UpdateSemesterWorkload(ctx context.Context, workload *Workload) error
 }
