+++ profileService/internal/service/profileVersion/service.go
@@ -19,7 +19,7 @@ func NewService(repo profileVersion.Repository, logger *zap.Logger) *Service {
 	return &Service{repo: repo, logger: logger}
 }
 
-func (s *Service) GetVersionByProfileID(ctx context.Context, profileID int64, year int) (*profileVersion.ProfileVersion, error) {
+func (s *Service) GetVersionByProfileID(ctx context.Context, profileID int64, year int64) (*profileVersion.ProfileVersion, error) {
 	return s.repo.GetVersionByProfileID(ctx, profileID, year)
 }
 
@@ -27,7 +27,7 @@ func (s *Service) GetVersionByVersionID(ctx context.Context, versionID int64) (*
 	return s.repo.GetVersionByVersionID(ctx, versionID)
 }
 
-func (s *Service) GetVersionIDByProfileID(ctx context.Context, profileID int64, year int) (int64, error) {
+func (s *Service) GetVersionIDByProfileID(ctx context.Context, profileID int64, year int64) (int64, error) {
 	version, err := s.GetVersionByProfileID(ctx, profileID, year)
 	if err != nil {
 		s.logger.Error("Failed to get profile version",
